+++ classes/output/courseformat/content/cm/delegatedcontrolmenu.php
@@ -16,23 +16,39 @@
 
 namespace format_flexsections\output\courseformat\content\cm;
 
-/**
- * Class delegatedcontrolmenu
- *
- * @package    format_flexsections
- * @copyright  Marina Glancy
- * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
- */
-class delegatedcontrolmenu extends \core_courseformat\output\local\content\cm\delegatedcontrolmenu {
+// The parent class only exists in Moodle 4.2+.
+if (class_exists(\core_courseformat\output\local\content\cm\delegatedcontrolmenu::class)) {
 
     /**
-     * Generate the edit control items of a section.
+     * Class delegatedcontrolmenu
      *
-     * @return array of edit control items
+     * @package    format_flexsections
+     * @copyright  Marina Glancy
+     * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
      */
-    public function delegated_control_items() {
-        $controls = parent::delegated_control_items();
-        unset($controls['view'], $controls['permalink']);
-        return $controls;
+    class delegatedcontrolmenu extends \core_courseformat\output\local\content\cm\delegatedcontrolmenu {
+
+        /**
+         * Generate the edit control items of a section.
+         *
+         * @return array of edit control items
+         */
+        public function delegated_control_items() {
+            $controls = parent::delegated_control_items();
+            unset($controls['view'], $controls['permalink']);
+            return $controls;
+        }
+    }
+} else {
+    // Stub to satisfy unittest (test_get_component_classes_in_namespace_provider).
+
+    /**
+     * Class delegatedcontrolmenu
+     *
+     * @package    format_flexsections
+     * @copyright  Marina Glancy
+     * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
+     */
+    class delegatedcontrolmenu {
     }
 }